requirements.txt
fastapi==0.100.0
gradio==4.29.0
gradio_client==0.16.1
httpcore==0.17.3
httpx==0.24.1
uvicorn==0.23.1
websockets==11.0.3
google-generativeai==0.5.4
ipython==8.25.0